requirements.txt
numpy
pandas
google-generativeai
python-dotenv
selenium
webdriver-manager
fastapi
uvicorn
httpx
pydantic